requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
decorator==4.4.2
futures==3.1.1
goslate==1.5.1
idna==2.10
lyricsgenius==2.0.2
pycparser==2.20
PyDictionary==2.0.1
pyshorteners==1.0.1
python-telegram-bot==13.0
pytube==9.7.0
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
tornado==6.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.11
wikipedia==1.4.0
